+++ concrete_datastore/api/v1/filters.py
@@ -106,9 +106,9 @@ def convert_type(string, field_type, close_period=True):
             # Expected format YYYY-MM-DD HH:MM[:ss[.uuuuuu]][TZ]
             dt = ensure_pendulum(string)
             if close_period:
-                dt = dt.end_of('day')
+                dt = dt.end_of('second')
             else:
-                dt = dt.start_of('day')
+                dt = dt.start_of('second')
             return format_datetime(dt)
         if field_type == 'DateField':
             # Expected YYYY-MM-DD